Add new test project(WSP authentication) and remove wrongly project references and code from CDP4Orm.Tests
