+++ CDP4-Server.sln
@@ -24,6 +24,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CDP4WebServer", "CDP4WebSer
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CDP4WspDatabaseAuthentication", "CDP4WSPDatabaseAuthentication\CDP4WspDatabaseAuthentication.csproj", "{E6245C47-96BA-40D3-9D5E-9114E175169F}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CDP4WspDatabaseAuthentication.Tests", "CDP4WspDatabaseAuthentication.Tests\CDP4WspDatabaseAuthentication.Tests.csproj", "{D95B61F1-4F95-4F7F-8C95-7B809A83FDEF}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -85,6 +87,12 @@ Global
 		{E6245C47-96BA-40D3-9D5E-9114E175169F}.Release|Any CPU.Build.0 = Release|Any CPU
 		{E6245C47-96BA-40D3-9D5E-9114E175169F}.Test|Any CPU.ActiveCfg = Test|Any CPU
 		{E6245C47-96BA-40D3-9D5E-9114E175169F}.Test|Any CPU.Build.0 = Test|Any CPU
+		{D95B61F1-4F95-4F7F-8C95-7B809A83FDEF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D95B61F1-4F95-4F7F-8C95-7B809A83FDEF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D95B61F1-4F95-4F7F-8C95-7B809A83FDEF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D95B61F1-4F95-4F7F-8C95-7B809A83FDEF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D95B61F1-4F95-4F7F-8C95-7B809A83FDEF}.Test|Any CPU.ActiveCfg = Test|Any CPU
+		{D95B61F1-4F95-4F7F-8C95-7B809A83FDEF}.Test|Any CPU.Build.0 = Test|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
